+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/authz/EntitlementAPITest.java
@@ -67,6 +67,7 @@ import org.keycloak.authorization.client.util.HttpResponseException;
 import org.keycloak.common.util.Base64Url;
 import org.keycloak.events.EventType;
 import org.keycloak.models.utils.KeycloakModelUtils;
+import org.keycloak.protocol.ProtocolMapperUtils;
 import org.keycloak.protocol.oidc.OIDCLoginProtocol;
 import org.keycloak.protocol.oidc.mappers.HardcodedClaim;
 import org.keycloak.protocol.oidc.mappers.OIDCAttributeMapperHelper;
@@ -1960,7 +1961,7 @@ public class EntitlementAPITest extends AbstractAuthzTest {
         customClaimMapper.setProtocol(OIDCLoginProtocol.LOGIN_PROTOCOL);
         Map<String, String> config = new HashMap<>();
         config.put(OIDCAttributeMapperHelper.TOKEN_CLAIM_NAME, "custom_claim");
-        config.put(HardcodedClaim.CLAIM_VALUE, PUBLIC_TEST_CLIENT);
+        config.put(ProtocolMapperUtils.CLAIM_VALUE, PUBLIC_TEST_CLIENT);
         config.put(OIDCAttributeMapperHelper.INCLUDE_IN_ACCESS_TOKEN, "true");
         customClaimMapper.setConfig(config);
 
@@ -1968,7 +1969,7 @@ public class EntitlementAPITest extends AbstractAuthzTest {
 
         ClientResource client = getClient(getRealm(), RESOURCE_SERVER_TEST);
 
-        config.put(HardcodedClaim.CLAIM_VALUE, RESOURCE_SERVER_TEST);
+        config.put(ProtocolMapperUtils.CLAIM_VALUE, RESOURCE_SERVER_TEST);
 
         client.getProtocolMappers().createMapper(customClaimMapper);