+++ CHANGELOG.md
@@ -36,7 +36,8 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 - Include claim in token introspection response only [RCIAM-742](https://jira.argo.grnet.gr/browse/RCIAM-742)
 - Enhanced TokenIntrospection and UserInfo events and logs
 - Configurable Claims for dynamic scopes, Filter dynamic scopes from access token scope [RCIAM-1190](https://jira.argo.grnet.gr/browse/RCIAM-1190)
-- Database script for migration from version 18.0.1-2.17 to 22.0.5-1.0
+- Database script for migration from version 18.0.1-2.17 to 22.0.5-1.1
+- Mapper for generating SAML attribute values or Claim values using IdP alias or User Attribute values [Mapper for generating SAML attribute values or Claim values using other SAML attribute/Claim values as input](https://trello.com/c/8K46f2mo/1642-mapper-for-generating-saml-attribute-values-or-claim-values-using-other-saml-attribute-claim-values-as-input)
 
 ### Changed
 - Increase User Attribute Value length to 4000 [EOSC-KC-132](https://github.com/eosc-kc/keycloak/issues/132)